Moved RecordedIntentsRule to testshared module and resued in PermissionDeniedDialogTest
