+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/settings/FormManagementSettingsTest.java
@@ -10,7 +10,6 @@ import org.junit.Test;
 import org.junit.rules.RuleChain;
 import org.junit.runner.RunWith;
 import org.odk.collect.android.R;
-import org.odk.collect.android.RecordedIntentsRule;
 import org.odk.collect.android.support.CollectTestRule;
 import org.odk.collect.android.support.AdbFormLoadingUtils;
 import org.odk.collect.android.support.NotificationDrawerRule;
@@ -20,6 +19,7 @@ import org.odk.collect.android.support.TestScheduler;
 import org.odk.collect.android.support.pages.FormManagementPage;
 import org.odk.collect.android.support.pages.ProjectSettingsPage;
 import org.odk.collect.android.support.pages.MainMenuPage;
+import org.odk.collect.testshared.RecordedIntentsRule;
 
 import java.util.List;
