+++ collect_app/src/androidTest/java/org/odk/collect/android/regression/ImageWidgetTest.kt
@@ -17,7 +17,6 @@ import org.junit.Test
 import org.junit.rules.RuleChain
 import org.odk.collect.android.BuildConfig
 import org.odk.collect.android.R
-import org.odk.collect.android.RecordedIntentsRule
 import org.odk.collect.android.application.Collect
 import org.odk.collect.android.storage.StoragePathProvider
 import org.odk.collect.android.support.CollectTestRule
@@ -25,6 +24,7 @@ import org.odk.collect.android.support.FileUtils
 import org.odk.collect.android.support.ResetStateRule
 import org.odk.collect.android.support.TestRuleChain
 import org.odk.collect.android.support.actions.NestedScrollToAction.nestedScrollTo
+import org.odk.collect.testshared.RecordedIntentsRule
 import java.io.File
 
 class ImageWidgetTest {
